# use fcopy for faster file copying