--- a/package.json
+++ b/package.json
@@ -15,6 +15,7 @@
     "death": "^1.0.0",
     "debug": "^2.2.0",
     "detect-indent": "^5.0.0",
+    "fcopy": "^0.0.7",
     "gunzip-maybe": "^1.4.0",
     "ini": "^1.3.4",
     "inquirer": "^3.0.1",