#### (this is not something that should be merged now, needs more testing)

I dug up my old uv_sendfile code and thought it could be interesting to make a little package out of it (https://github.com/sciolist/fcopy) and see what impact it would have on yarn.

### My test setup:
  - yarn in offline mode
  - yarn.lock exists, node_modules is empty
  - windows defender disabled
  - node v7.8.0
  - [package.json with about 100 deps, 42k files, 200mb](https://github.com/sciolist/fcopy/blob/c0ded8b82bbaa3f4cdb0a67e951eb3f8b0f97fb7/perf/project/alottafiles/package.json)
  - each test ran 10 times

### Results

| Platform | avg speed v0.24 (master) | avg speed with fcopy lib |
| - | - | - |
| FreeBSD | 32s | 9s |
| Ubuntu | 30s | 9s |
| Os x | 32s | 10s |
| Windows | 68s | 32s |

I've ran diffs on the output and worked with it a bit today without noticing any differences or issues, but, I wouldn't be surprised if there are some.. particularly on other bsds.