@Daniel15 2017-05-12
Probably worth collaborating with @vbfox on this - He was looking at using the Win32 CopyFile API for copying files (#2960). Perhaps the Win32 stuff could be merged into `fcopy`.