@vbfox 2017-05-12
So I tried on win32 with fcopy vs win32 CopyFile and there is no big gain to be made (**~1%** on the test project) but there is quite a bit of gain possible on win32 by removing the `utimes` call if we're sure that we're calling the native `CopyFile` (Windows preserve `mtime` by default and it's the only field Yarn seem to use).

On my PC this PR take **11.75s** and without the `utime` call it's **10.21s**