@sciolist 2017-05-12
So I've been using it for about 2 weeks now, on os x and windows, a bit on linux.. without any issues so far.

Based on the perf. reported in https://github.com/yarnpkg/yarn/pull/2960 it seems like this performs similarly to FileCopy, but a fair bit of that would probably be from ffi. I'll be interested in seeing if he finds a way to speed it up for windows.

One problem is that you need a version of the native extension built for your major version of node (specifically NODE_MODULE_VERSION from https://nodejs.org/en/download/releases/.) That makes distribution a bit tricky.

If you're using a mismatching node version it'll fall back to readstream+writestream just like yarn copies files today, so it isn't the end of the world.. and if you want to keep distributing yarn as a single js file that'll work fine as well.

My current setup is that it'll download the correct extension from https://github.com/sciolist/fcopy/releases/tag/v0.0.5 for your version of node when you `yarn add fcopy`.. But that of course won't work if you're installing yarn through an .msi. One approach would be to just download the binaries for all supported node versions up front so it could pick the correct one at runtime. Another option I was considering is to download the version it needs at runtime, but, that just doesn't feel very good.