@vbfox 2017-05-12
I think packaging the binary with yarn would be nice. The total size per platform over supported node versions don't seem so big & the fallback is always there if the user got another version or is using the single-file version.

Also in the end I checked and `mtime` can't be optimized aways when `fcopy` is used as it's necessary here (libuv does the copy manually not by using OS functions so it doesn't use win32 API)
I don't think a win32 copier in `fcopy` is really needed (1% speedup isn't really worth it)  but if you want I can PR it (The code is there if needed : https://github.com/vbfox/native-copy-file/blob/gyp/src/hello.cc)

My times btw

| 0.23.2        | CopyFile           |  libuv  |
| ------------- |:-------------:| -----:|
| 24.97s      | 11.48s | 11.75s |
| 100%      | 46% | 47% |