@Daniel15 2017-05-12
> My current setup is that it'll download the correct extension from https://github.com/sciolist/fcopy/releases/tag/v0.0.5 for your version of node when you yarn add fcopy.. But that of course won't work if you're installing yarn through an .msi. One approach would be to just download the binaries for all supported node versions up front so it could pick the correct one at runtime. 

The binaries look pretty small, we could probably bundle all of them with Yarn and just load the correct one at runtime, if possible. I don't know a lot about how native extensions work though, and whether they can be dynamically loaded.