@bestander 2017-05-12
Silly question.
Is it possible to do it without Native code?
From distribution point of view, adding extra binary make it more complex


On Fri, 12 May 2017 at 23:36, Daniel Lo Nigro <notifications@github.com>
wrote:

> My current setup is that it'll download the correct extension from
> https://github.com/sciolist/fcopy/releases/tag/v0.0.5 for your version of
> node when you yarn add fcopy.. But that of course won't work if you're
> installing yarn through an .msi. One approach would be to just download the
> binaries for all supported node versions up front so it could pick the
> correct one at runtime.
>
> The binaries look pretty small, we could probably bundle all of them with
> Yarn and just load the correct one at runtime, if possible. I don't know a
> lot about how native extensions work though, and whether they can be
> dynamically loaded.
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/yarnpkg/yarn/pull/3290#issuecomment-301201166>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/ACBdWNFfy6HAd_fyMI64jb34Holgg6oqks5r5N7_gaJpZM4NMvTC>
> .
>