@sciolist 2017-05-12
@Daniel15 Yeah you can dynamically load them, you just `require()` the built .node file and it'll throw an exception if you try to load an incompatible library.. I could probably just package all the different versions into one tar.gz file per platform instead, that way no changes would have to be made in yarn.

@vbfox for a 1% gain it feels like it's probably preferable to just stick with libuv. :/ a bit less code to maintain, and the libuv code is probably at least somewhat battle tested. regarding utimes, you could maybe skip actually `await`-ing the utimes result.. not sure if that would cause other issues.