@sciolist 2017-05-12
@bestander not really, though if fcopys post-install downloads binaries for all supported node versions there shouldn't have to be more handling in yarn than for any other library. the alternative would be to convince the node maintainers that a fast file copy function should be added to `fs`.. not sure how receptive they'd be to that.