@sciolist 2017-05-14
@Daniel15 Made a few updates, `yarn add fcopy` downloads binaries for several v8 versions.. I'll have to try out yarns dist builds to see how well that will work with them.

The function for downloading binaries is exposed now, so if there is no local version of the binary it can attempt to download it:
```js
require('fcopy/lib/bindings').downloadIfNeeded(process.versions.modules).then(
  function () {},
  function (ex) { /* will fall back to fs streams */ }
);
```
Not sure if it should actually be used though.. For most situations it doesn't feel great to just download and execute binaries off the internet, but then yarn already does that, so maybe not a big deal.

I figured I'd also put up a quick video comparison of the current file copying in yarn:

[fs streams (osx)](https://asciinema.org/a/8y8sx5yjdespk8qqqmekmzzci), taking 9.3 seconds to copy 42k files (304M)

[fcopy (osx)](https://asciinema.org/a/8jk9ra0daujxrdbp654re0qea), taking 2.4 seconds for the same files.

It is a pretty significant difference, but if it's worth adding a binary dependency to get it is hard for me to say.