@Daniel15 2017-06-06
This will be useful for copy-on-write filesystems and should provide a significant speed improvement for systems that use CoW filesystems. It would be good to get this in, we just need to think about the deployment strategy and how to handle it for the single-file build. Falling back to the JS code when the native library isn't available is probably sufficient.