@Daniel15 2017-06-06
On the other hand, maybe simply generating a shell script with all the required `cp --reflink=auto` commands and execing that would be more maintainable than native code, as Node.js' handling of native modules is notoriously bad. 😛