@sciolist 2017-06-06
Getting a copyfile function in libuv+node would be great, though I'm not sure what the likelihood is of that happening particularly soon.

Creating a shell script does have a compatibility advantage over a native module, since as you said they're pretty painful to maintain.. The biggest problem it that changing your local node version stops the native module from working, unless you keep a lot of different versions of the native module around (and when a new version of node comes out, you'd need to download a new native module.) I do have a fallback for that, but it isn't great that changing node version substantially reduces copy speed.

Definitely worth looking into alternatives.